Giant refactor 1
The new program to launch both services as part of the same EventMachine
reactor is webapp.rb. Future versions may involve a refactor to a
Rack-based interface. The configuration file format has been cleaned up
and pulled in in a few places. This is confirmed functionally identical
to the existing website.